Fixed the type of variable 'n_term' in the documentation

+++ fibonacci.cpp
@@ -20,14 +20,15 @@ using namespace std;
     The function **fibonacci_binet_formula** returns the n'th term of the fibonacci sequence
 
     @param n_term
-        - ** Integer (int) ** - term of fibonacci sequence that will be calculated
+        - ** unsigned int ** - term of fibonacci sequence that will be calculated
+
     @return 
-        - ** double (double) ** - n'th term of the fibonacci sequence
+        - ** double ** - n'th term of the fibonacci sequence
 ***************************************************/
 
 double fibonacci_binet_formula(unsigned int n_term){
 
     // To calculate the n'th term of the fibonacci sequencie was use the Binet's Formula
-    double n_term_value = (pow(((1 + sqrt(5))/2),n_term) - pow(((1 - sqrt(5))/2), n_term))/sqrt(5);
+    double n_term_value = (pow(((1 + sqrt(5))/2), n_term) - pow(((1 - sqrt(5))/2), n_term))/sqrt(5);
     return n_term_value;
 }
